% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_expranno_data}
\alias{example_expranno_data}
\title{Small built-in example inputs}
\usage{
example_expranno_data(species = c("human", "mouse"))
}
\arguments{
\item{species}{Either \code{"human"} or \code{"mouse"}.}
}
\value{
A list with \code{expr} and \code{meta}.
}
\description{
Small built-in example inputs
}
