% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{read_genesets}
\alias{read_genesets}
\title{Read gene sets from a GMT file}
\usage{
read_genesets(geneset_file)
}
\arguments{
\item{geneset_file}{A \code{.gmt} file path.}
}
\value{
A named list of character vectors.
}
\description{
Read gene sets from a GMT file
}
