% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_expr}
\alias{validate_expr}
\title{Validate an expression table}
\usage{
validate_expr(expr)
}
\arguments{
\item{expr}{A data frame whose first column is \code{gene_id} and remaining
columns are numeric sample columns.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Validate an expression table
}
