% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_meta}
\alias{validate_meta}
\title{Validate a metadata table}
\usage{
validate_meta(meta)
}
\arguments{
\item{meta}{A data frame whose first column is \code{sample}.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Validate a metadata table
}
