% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{read_genesets}
\alias{read_genesets}
\title{Read gene sets}
\usage{
read_genesets(x)
}
\arguments{
\item{x}{A named list of gene sets or a path to a GMT file.}
}
\value{
A named list of character vectors.
}
\description{
Read gene sets
}
