% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{validate_meta}
\alias{validate_meta}
\title{Validate a metadata table}
\usage{
validate_meta(meta)
}
\arguments{
\item{meta}{A sample metadata table.}
}
\value{
\code{TRUE} invisibly on success.
}
\description{
\code{meta} must be a \code{data.frame} with \code{sample} as the first column and one row
per sample.
}
